Fix OSS to use a buffer of shorts and not floats
git-svn-id: https://svn.code.sf.net/p/sidplay-residfp/code/trunk/sidplayfp@840 75748be7-0387-4326-94c7-f41ce0883d2c

+++ src/audio/oss/audiodrv.cpp
@@ -141,9 +141,9 @@ short *Audio_OSS::open (AudioConfig &cfg, const char *)
     ioctl (_audiofd, SNDCTL_DSP_GETBLKSIZE, &temp);
     cfg.bufSize = (uint_least32_t) temp;
 #ifdef HAVE_EXCEPTIONS
-    _sampleBuffer = new(std::nothrow) float[cfg.bufSize];
+    _sampleBuffer = new(std::nothrow) short[cfg.bufSize];
 #else
-    _sampleBuffer = new float[cfg.bufSize];
+    _sampleBuffer = new short[cfg.bufSize];
 #endif
 
     if (!_sampleBuffer)
